refactor(ai_contracts): de-part selection, split near-cap libs, shared H8 fixtures (Refs #4084)
Convert the Full AI civilian-work part cluster and simple-heuristics splits into
explicit-import libraries, topic-split OW feedstock localization, densify H8
seller/supplier test builders under test/support/, and add
repo.ai_contracts_no_part_directives.
