+++ SPEC/program/repo-lint.md
@@ -154,7 +154,8 @@
 | `tool/check_combat_test_scenario_harness.dart` | Forbid bare `for (final scenario in …)` loops under `packages/colonizethis_combat/test/**/*_test.dart`; scenario tables must register via `runLabeledScenarios` / `runLabeledScenarioGroup` from `colonizethis_combat_test_support` (Refs #3983); rule `repo.combat_test_scenario_harness` |
 | `tool/check_economy_no_part_directives.dart` | Forbid Dart `part` / `part of` directives under `packages/colonizethis_economy/lib/**` so DealMatcher matching/indexing sub-files stay proper libraries with explicit imports (mirrors world/turn/diplomacy/orders no-part gates; Refs #3979); rule `repo.economy_no_part_directives` |
 | `tool/check_models_no_part_directives.dart` | Forbid Dart `part` / `part of` directives under `packages/colonizethis_models/lib/**` so orders/world_state/world_market/app_events sub-files stay proper libraries with explicit imports (mirrors economy/world/turn no-part gates; Refs #4068); rule `repo.models_no_part_directives` |
-| `tool/check_ai_no_part_directives.dart` | Forbid Dart `part` / `part of` directives under `packages/colonizethis_ai/lib/**` so planning/perception sub-files stay proper libraries with explicit imports (mirrors economy/world/turn/models no-part gates; Refs #4079); rule `repo.ai_no_part_directives` |
+| `tool/check_ai_no_part_directives.dart` | Forbid Dart `part` / `part of` directives under `packages/colonizethis_ai/lib/**` so planning/perception sub-files stay proper libraries with explicit imports (mirrors economy/world/turn/models no-part gates; Refs #4079); rule `repo.ai_no_part_directives`. Also hosts the shared parameterized scanner used by `repo.ai_contracts_no_part_directives` (Refs #4084). |
+| `tool/check_ai_contracts_no_part_directives.dart` | Forbid Dart `part` / `part of` directives under `packages/colonizethis_ai_contracts/lib/**` so Full AI civilian selection / heuristics / localization sub-files stay proper libraries with explicit imports (mirrors `repo.ai_no_part_directives`; Refs #4084) — see `SPEC/program/ai-contracts-no-part-directives.md`; rule `repo.ai_contracts_no_part_directives` |
 | `tool/check_ai_source_file_size.dart` | Enforce `packages/colonizethis_ai/lib/src/**` Dart files at **500 physical lines or fewer**, with an optional shrink-only grandfather allowlist that fails when entries are missing or now under-cap (mirrors save/data size gates; ceiling matches domain-package source size; Refs #4079 Slice C) — see `SPEC/program/ai-source-file-size.md`; rule `repo.ai_source_file_size` |
 | `tool/check_models_test_file_size.dart` | Fail when any `.dart` file under `packages/colonizethis_models/test/**` exceeds 400 physical lines (aligns with diplomacy/economy test file caps; Refs #4068 Slice D); rule `repo.models_test_file_size` |
 | `tool/check_world_topology_identity_home.dart` | Require topology identity helpers (`indexTopologyNodesByRegion`, `provinceTopologyNodeId`, `provinceIdsAdjacentToSeaZone`, `regionIdForSeaZone`) to be defined in `packages/colonizethis_world/lib/src/world/topology_identity.dart` and forbid re-declaring them as top-level functions in `naval.dart` (re-export allowed; Refs #3968); rule `repo.world_topology_identity_home` |
@@ -503,6 +504,8 @@ Do **not** add new top-level `tool/check_*.dart` **entrypoints** for CI without 
 - Given a `packages/colonizethis_models/lib/**` Dart file that contains `part 'foo.dart';` or `part of 'foo.dart';`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.models_no_part_directives`, then the run fails listing each offending `path:line` and exits `1` (Refs #4068).
 - Given the `colonizethis_ai` package on `dev` where no `lib/**` file uses a `part` / `part of` directive, when `dart run tool/ct_repo_lint.dart` runs rule `repo.ai_no_part_directives`, then the rule passes without violations and exits `0` (Refs #4079).
 - Given a `packages/colonizethis_ai/lib/**` Dart file that contains `part 'foo.dart';` or `part of 'foo.dart';`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.ai_no_part_directives`, then the run fails listing each offending `path:line` and exits `1` (Refs #4079).
+- Given the `colonizethis_ai_contracts` package on `dev` where no `lib/**` file uses a `part` / `part of` directive, when `dart run tool/ct_repo_lint.dart` runs rule `repo.ai_contracts_no_part_directives`, then the rule passes without violations and exits `0` (Refs #4084).
+- Given a `packages/colonizethis_ai_contracts/lib/**` Dart file that contains `part 'foo.dart';` or `part of 'foo.dart';`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.ai_contracts_no_part_directives`, then the run fails listing each offending `path:line` and exits `1` (Refs #4084).
 - Given every hand-written file under `packages/colonizethis_ai/lib/src/**` is at or below 500 physical lines and the grandfather allowlist is empty, when `dart run tool/ct_repo_lint.dart` runs rule `repo.ai_source_file_size`, then the rule passes without violations and exits `0` (Refs #4079).
 - Given a hand-written `packages/colonizethis_ai/lib/src/**` Dart file with more than 500 physical lines and an empty grandfather allowlist, when `dart run tool/ct_repo_lint.dart` runs rule `repo.ai_source_file_size`, then the run fails naming that file and exits `1` (Refs #4079).
 - Given a grandfather allowlist entry for `repo.ai_source_file_size` that names a missing file or a file now at or below 500 physical lines, when the rule runs, then the run fails reporting a stale grandfather entry and exits `1` (Refs #4079).
