+++ packages/colonizethis_ai_contracts/lib/src/ai/full_ai_civilian_work_selection_build_purchase.dart
@@ -1,10 +1,18 @@
-part of 'full_ai_civilian_work_selection.dart';
+import 'package:colonizethis_data/colonizethis_data.dart';
+import 'package:colonizethis_models/colonizethis_models.dart';
+
+import '../constants.dart';
+import 'full_ai_civilian_work_selection_feedstock_predicates.dart';
+import 'full_ai_civilian_work_selection_shared.dart';
+import 'package:colonizethis_world/src/world/faction_membership.dart';
+import 'package:colonizethis_world/src/world/player_view.dart';
+import 'package:colonizethis_world/src/world/province_lookup.dart';
 
 // Builder (`build_improvement`) and Merchant (`purchase_land`) candidate
 // scoring / row selection, plus the Old World feedstock unit reservation that
 // keeps an idle Builder / Explorer available for H8 feedstock work. Split out
 // of full_ai_civilian_work_selection.dart by concern to keep each library file
-// small; shares the parent library's private scope via `part`.
+// small.
 
 // The civilian-work feedstock score boosts
 // ([kRegimentBuildInputFeedstockExtractionScoreBoost],
@@ -13,7 +21,7 @@ part of 'full_ai_civilian_work_selection.dart';
 // `ai_victory_config.dart` / `ai_parameter_registry.dart` (Refs #3794);
 // behaviour is normative in SPEC/ai/civilian-work-planner.md.
 
-int _buildImprovementWorkScore(
+int buildImprovementWorkScore(
   WorkOrder w,
   Game game, {
   required String playerId,
@@ -56,54 +64,54 @@ int _buildImprovementWorkScore(
 /// The seed-42 affluent suppliers own an unimproved Old World `iron` / `timber`
 /// feedstock tile on every gate-active turn, yet every idle Builder / Explorer
 /// is routed to higher-scoring **New World** owned-resource work (the New World
-/// bonuses in [_buildImprovementWorkScore] / [_eScore]), so the now-correct
+/// bonuses in [buildImprovementWorkScore] / explore scoring), so the now-correct
 /// prospecting / co-availability ordering chain never gets a unit positioned on
 /// the Old World feedstock tile. Holding one idle Builder and one idle Explorer
 /// out of New World work keeps them available for the Old World feedstock
 /// `prospect` + `build_improvement` the existing
 /// [kFeedstockMineralProspectScoreBoost] /
 /// [kRegimentBuildInputFeedstockExtractionScoreBoost] boosts then select.
-class _OwFeedstockReservation {
-  const _OwFeedstockReservation({this.builderUnitId, this.explorerUnitId});
+class OwFeedstockReservation {
+  const OwFeedstockReservation({this.builderUnitId, this.explorerUnitId});
 
   final String? builderUnitId;
   final String? explorerUnitId;
 
-  static const none = _OwFeedstockReservation();
+  static const none = OwFeedstockReservation();
 
   bool reserves(String unitId) =>
       unitId == builderUnitId || unitId == explorerUnitId;
 }
 
-/// Resolves the [_OwFeedstockReservation] for the active player.
+/// Resolves the [OwFeedstockReservation] for the active player.
 ///
-/// Returns [_OwFeedstockReservation.none] unless the feedstock-extraction gate
+/// Returns [OwFeedstockReservation.none] unless the feedstock-extraction gate
 /// is active ([feedstockExtractionResourceIds] non-empty). When active, reserves
 /// the lexicographically-smallest idle (`currentWork == null`) Builder iff the
 /// player owns an unimproved Old World feedstock tile, and the
 /// lexicographically-smallest idle Explorer iff the player owns an unprospected
 /// Old World mineral feedstock tile. Deterministic over
 /// `(view.ownUnits, game, feedstockExtractionResourceIds)`.
-_OwFeedstockReservation _resolveOwFeedstockReservation(
+OwFeedstockReservation resolveOwFeedstockReservation(
   PlayerView view,
   Game game,
   Set<String> feedstockExtractionResourceIds,
 ) {
   if (feedstockExtractionResourceIds.isEmpty) {
-    return _OwFeedstockReservation.none;
+    return OwFeedstockReservation.none;
   }
   final playerId = view.playerId;
-  final reserveBuilder = _ownsUnimprovedOldWorldFeedstockTile(
+  final reserveBuilder = ownsUnimprovedOldWorldFeedstockTile(
     game,
     playerId,
     feedstockExtractionResourceIds,
   );
-  final reserveExplorer = _ownsUnprospectedOldWorldMineralFeedstockTile(
+  final reserveExplorer = ownsUnprospectedOldWorldMineralFeedstockTile(
     game,
     playerId,
     feedstockExtractionResourceIds,
   );
-  if (!reserveBuilder && !reserveExplorer) return _OwFeedstockReservation.none;
+  if (!reserveBuilder && !reserveExplorer) return OwFeedstockReservation.none;
   final idleBuilders = <String>[];
   final idleExplorers = <String>[];
   for (final unit in view.ownUnits) {
@@ -117,7 +125,7 @@ _OwFeedstockReservation _resolveOwFeedstockReservation(
   }
   idleBuilders.sort();
   idleExplorers.sort();
-  return _OwFeedstockReservation(
+  return OwFeedstockReservation(
     builderUnitId: idleBuilders.isEmpty ? null : idleBuilders.first,
     explorerUnitId: idleExplorers.isEmpty ? null : idleExplorers.first,
   );
@@ -128,13 +136,13 @@ _OwFeedstockReservation _resolveOwFeedstockReservation(
 /// unit with only its Old World candidates (or none, in which case it stays idle
 /// in the Old World). Refs #2847 § H8-extraction supplier Old World feedstock
 /// unit reservation.
-void _dropNewWorldWorkOrders(List<WorkOrder> orders) {
+void dropNewWorldWorkOrders(List<WorkOrder> orders) {
   orders.removeWhere(
     (w) => Unit.regionIdFromTileKey(w.targetTileKey) == kNewWorldRegionId,
   );
 }
 
-WorkOrder? _bestBuildImprovementRow(
+WorkOrder? bestBuildImprovementRow(
   List<WorkOrder> candidates,
   Game game, {
   required String playerId,
@@ -147,7 +155,7 @@ WorkOrder? _bestBuildImprovementRow(
       .toList();
   if (improvements.isEmpty) return null;
   var best = improvements.first;
-  var bestScore = _buildImprovementWorkScore(
+  var bestScore = buildImprovementWorkScore(
     best,
     game,
     playerId: playerId,
@@ -158,7 +166,7 @@ WorkOrder? _bestBuildImprovementRow(
   );
   for (var i = 1; i < improvements.length; i++) {
     final w = improvements[i];
-    final s = _buildImprovementWorkScore(
+    final s = buildImprovementWorkScore(
       w,
       game,
       playerId: playerId,
@@ -173,7 +181,7 @@ WorkOrder? _bestBuildImprovementRow(
       best = w;
       continue;
     }
-    if (s == bestScore && _compareWorkOrderLex(w, best) < 0) {
+    if (s == bestScore && compareWorkOrderLex(w, best) < 0) {
       best = w;
     }
   }
@@ -199,7 +207,7 @@ int _purchaseLandWorkScore(
   return 60;
 }
 
-WorkOrder? _bestPurchaseLandRow(
+WorkOrder? bestPurchaseLandRow(
   List<WorkOrder> candidates,
   Game game,
   DiplomacyFactionMembership factionMembership,
@@ -218,7 +226,7 @@ WorkOrder? _bestPurchaseLandRow(
       best = w;
       continue;
     }
-    if (s == bestScore && _compareWorkOrderLex(w, best) < 0) {
+    if (s == bestScore && compareWorkOrderLex(w, best) < 0) {
       best = w;
     }
   }
