+++ packages/colonizethis_ai_contracts/lib/src/ai/full_ai_civilian_work_selection_engineer.dart
@@ -1,11 +1,16 @@
-part of 'full_ai_civilian_work_selection.dart';
+import 'package:colonizethis_data/colonizethis_data.dart';
+import 'package:colonizethis_models/colonizethis_models.dart';
+
+import '../constants.dart';
+import 'full_ai_civilian_work_selection.dart' show FullAiCivilianWorkIdle;
+import 'full_ai_civilian_work_selection_shared.dart';
 
 // Engineer (`build_road` / `build_port` / `build_fort`) candidate scoring / row
 // selection and the Engineer path appender. Replaces the lexicographic fallback
 // (which always picked `build_fort` first because it sorts alphabetically before
 // `build_port`/`build_road`) with a unified scored pool over all three Engineer
 // targets. Split out of full_ai_civilian_work_selection.dart by concern to keep
-// each library file small; shares the parent library's private scope via `part`.
+// each library file small.
 //
 // Normative SPEC: SPEC/ai/civilian-work-planner.md § Engineer (Refs #3794).
 // Per-target base weights express the relative priority of the three targets;
@@ -101,7 +106,7 @@ WorkOrder? _bestEngineerRow(
   return best;
 }
 
-void _appendEngineerPathResult({
+void appendEngineerPathResult({
   required Unit? unit,
   required List<WorkOrder> w,
   required Game game,
@@ -110,7 +115,7 @@ void _appendEngineerPathResult({
   required List<FullAiCivilianWorkIdle> idleEvents,
 }) {
   final chosen =
-      _bestEngineerRow(w, game, playerId: playerId) ?? _pickLexicographic(w);
+      _bestEngineerRow(w, game, playerId: playerId) ?? pickLexicographic(w);
   if (chosen != null) {
     workOrders.add(chosen);
     return;
