+++ packages/colonizethis_ai_contracts/lib/src/ai/full_ai_civilian_work_selection_rail.dart
@@ -1,11 +1,16 @@
-part of 'full_ai_civilian_work_selection.dart';
+import 'package:colonizethis_data/colonizethis_data.dart';
+import 'package:colonizethis_models/colonizethis_models.dart';
+
+import '../constants.dart';
+import 'full_ai_civilian_work_selection.dart' show FullAiCivilianWorkIdle;
+import 'full_ai_civilian_work_selection_shared.dart';
 
 // Rail Builder (`build_rail`) candidate scoring / row selection and the Rail
 // Builder path appender. Replaces the lexicographic fallback (which always
 // picked the alphabetically-smallest tile) with a unified scored pool over the
 // unit's `build_rail` candidates. Split out of
 // full_ai_civilian_work_selection.dart by concern to keep each library file
-// small; shares the parent library's private scope via `part`.
+// small.
 //
 // Normative SPEC: SPEC/ai/civilian-work-planner.md § Rail Builder (Refs #3794
 // AC6). Scoring factors (all GA-tunable via ai_victory_config.dart):
@@ -81,7 +86,7 @@ WorkOrder? _bestBuildRailRow(
   return best;
 }
 
-void _appendRailBuilderPathResult({
+void appendRailBuilderPathResult({
   required Unit? unit,
   required List<WorkOrder> w,
   required Game game,
@@ -90,7 +95,7 @@ void _appendRailBuilderPathResult({
   required List<FullAiCivilianWorkIdle> idleEvents,
 }) {
   final chosen =
-      _bestBuildRailRow(w, game, playerId: playerId) ?? _pickLexicographic(w);
+      _bestBuildRailRow(w, game, playerId: playerId) ?? pickLexicographic(w);
   if (chosen != null) {
     workOrders.add(chosen);
     return;
