+++ packages/colonizethis_ai_contracts/lib/src/ai/full_ai_civilian_work_selection_spy.dart
@@ -1,4 +1,10 @@
-part of 'full_ai_civilian_work_selection.dart';
+import 'package:colonizethis_data/colonizethis_data.dart';
+import 'package:colonizethis_models/colonizethis_models.dart';
+
+import '../constants.dart';
+import 'full_ai_civilian_work_selection.dart' show FullAiCivilianWorkIdle;
+import 'full_ai_civilian_work_selection_shared.dart';
+import 'package:colonizethis_world/src/world/unit_lookup.dart';
 
 // Spy (`counter_spy`) candidate scoring / row selection. Refs #3834 R11.
 // Empire-wide counter-espionage: one spy on counter_spy anywhere is sufficient.
@@ -104,7 +110,7 @@ WorkOrder? _bestSpyRow(
   return best;
 }
 
-void _appendSpyPathResult({
+void appendSpyPathResult({
   required Unit? unit,
   required List<WorkOrder> w,
   required Game game,
@@ -125,7 +131,7 @@ void _appendSpyPathResult({
         enemySpyProvinceIds: enemySpyProvinceIds,
         playerAlreadyHasCounterSpy: playerAlreadyHasCounterSpy,
       ) ??
-      _pickLexicographic(w);
+      pickLexicographic(w);
   if (chosen != null) {
     workOrders.add(chosen);
     return;
