+++ packages/colonizethis_ai_contracts/lib/src/ai/full_ai_civilian_work_selection_upgrade_town.dart
@@ -1,4 +1,8 @@
-part of 'full_ai_civilian_work_selection.dart';
+import 'package:colonizethis_data/colonizethis_data.dart';
+import 'package:colonizethis_models/colonizethis_models.dart';
+
+import '../constants.dart';
+import 'full_ai_civilian_work_selection_build_purchase.dart';
 
 // Builder `upgrade_town` candidate scoring / row selection. Adds the
 // `upgrade_town` target to the Builder's scored pool so town upgrades compete
@@ -6,7 +10,7 @@ part of 'full_ai_civilian_work_selection.dart';
 // the lexicographic fallback (which picked an `upgrade_town` candidate only when
 // no `build_improvement` candidate existed). Split out of
 // full_ai_civilian_work_selection.dart by concern to keep each library file
-// small; shares the parent library's private scope via `part`.
+// small.
 //
 // Normative SPEC: SPEC/ai/civilian-work-planner.md § Builder (Refs #3794).
 // `upgrade_town` scoring is a GA-tunable baseline plus context bonuses using the
@@ -98,7 +102,7 @@ int _compareBuilderCrossType(WorkOrder a, WorkOrder b) {
 /// scored pool. The highest-scoring candidate across **both** target types wins;
 /// when no `upgrade_town` candidate exists the result is byte-identical to the
 /// `build_improvement`-only selection (Refs #3794 § Builder, no-regression).
-WorkOrder? _bestBuilderRow(
+WorkOrder? bestBuilderRow(
   List<WorkOrder> candidates,
   Game game, {
   required String playerId,
@@ -106,7 +110,7 @@ WorkOrder? _bestBuilderRow(
   Set<String> growthStageFabricFeedstockResourceIds = const <String>{},
   Set<String> growthStageInfraFeedstockResourceIds = const <String>{},
 }) {
-  final improvement = _bestBuildImprovementRow(
+  final improvement = bestBuildImprovementRow(
     candidates,
     game,
     playerId: playerId,
@@ -117,7 +121,7 @@ WorkOrder? _bestBuilderRow(
   final upgrade = _bestUpgradeTownRow(candidates, game, playerId: playerId);
   if (improvement == null) return upgrade;
   if (upgrade == null) return improvement;
-  final iScore = _buildImprovementWorkScore(
+  final iScore = buildImprovementWorkScore(
     improvement,
     game,
     playerId: playerId,
