+++ packages/colonizethis_ai_contracts/test/full_ai_civilian_work_ow_feedstock_prospect_mineral_eligibility_test.dart
@@ -1,10 +1,11 @@
-import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_ai_contracts/colonizethis_ai_contracts.dart';
+import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_logic/colonizethis_logic.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_test/test.dart';
 
-import 'full_ai_civilian_work_supplier_feedstock_extraction_fixtures.dart';
+import 'support/full_ai_civilian_work_supplier_feedstock_extraction_fixtures.dart';
+import 'support/h8_supplier_prospect_game.dart';
 
 // Refs #2847 § H8-extraction Old World mineral feedstock prospect localization
 // (terrain mineral-eligibility gate slice).
@@ -17,46 +18,11 @@ import 'full_ai_civilian_work_supplier_feedstock_extraction_fixtures.dart';
 // hills/mountain/swamp/desert). This predicate adds the live-terrain
 // mineral-eligibility check on top of the resource-only co-location signal.
 
-const _supplierIronTile = 'oldWorld|s0|2|0';
-const _supplierTimberTile = 'oldWorld|s0|1|0';
-const _newWorldIronTile = 'newWorld|n0|0|0';
-
-Game _supplierGame({
-  Map<String, String>? resourceByTileKey,
-  List<Unit> extraUnits = const [],
-}) {
-  return twoPlayerSupplierFeedstockGame(
-    resourceByTileKey:
-        resourceByTileKey ??
-        const {
-          _supplierTimberTile: 'timber',
-          _supplierIronTile: 'iron',
-          sellerWoolTile: 'wool',
-        },
-    extraUnits: extraUnits,
-  );
-}
-
-Unit _explorer(String id, {CurrentWork? currentWork}) =>
-    _explorerAt(id, 'oldWorld|s0', currentWork: currentWork);
-
-Unit _explorerAt(
-  String id,
-  String locationProvinceId, {
-  CurrentWork? currentWork,
-}) => Unit(
-  id: id,
-  type: kUnitTypeExplorer,
-  ownerId: supplierFeedstockId,
-  locationProvinceId: locationProvinceId,
-  currentWork: currentWork,
-);
-
 // The supplier `iron` tile sits at (x=2, y=0) of region `oldWorld`. This live
 // tile map sets the iron tile's terrain to [ironTerrain] so the mineral-
 // eligibility terrain check can pass (hills/mountain/swamp/desert) or fail
 // (plains/forest); the other cells are hills (prospectable filler).
-Map<String, TileMapResult> _tileMapWithIronTerrain(TerrainType ironTerrain) => {
+Map<String, TileMapResult> tileMapWithIronTerrain(TerrainType ironTerrain) => {
   'oldWorld': TileMapResult(
     width: 3,
     height: 1,
@@ -76,13 +42,13 @@ void main() {
     () {
       test('true when the co-located iron tile is on prospectable terrain '
           '(hills)', () {
-        final game = _supplierGame(extraUnits: [_explorer('e1')]);
+        final game = supplierGame(extraUnits: [supplierProspectExplorer('e1')]);
         expect(
           ownsIdleExplorerColocatedWithMineralEligibleUnprospectedOldWorldFeedstockTile(
             game,
             supplierFeedstockId,
             {'iron'},
-            _tileMapWithIronTerrain(TerrainType.hills),
+            tileMapWithIronTerrain(TerrainType.hills),
           ),
           isTrue,
         );
@@ -91,7 +57,7 @@ void main() {
       test('false when the co-located iron tile is on non-prospectable terrain '
           '(plains) — localizes the residual to terrain mineral-eligibility',
           () {
-        final game = _supplierGame(extraUnits: [_explorer('e1')]);
+        final game = supplierGame(extraUnits: [supplierProspectExplorer('e1')]);
         // Teeth: the resource-only co-located predicate is still true (the
         // supplier owns an idle Explorer on the unprospected iron province),
         // so the only discriminator is the live-terrain mineral-eligibility
@@ -109,7 +75,7 @@ void main() {
             game,
             supplierFeedstockId,
             {'iron'},
-            _tileMapWithIronTerrain(TerrainType.plains),
+            tileMapWithIronTerrain(TerrainType.plains),
           ),
           isFalse,
         );
@@ -121,7 +87,7 @@ void main() {
         // resource-only signal, matching `isMineralEligibleTile`'s documented
         // missing-terrain fallback, so the predicate stays equivalent to the
         // resource-only co-located counterpart.
-        final game = _supplierGame(extraUnits: [_explorer('e1')]);
+        final game = supplierGame(extraUnits: [supplierProspectExplorer('e1')]);
         expect(
           ownsIdleExplorerColocatedWithMineralEligibleUnprospectedOldWorldFeedstockTile(
             game,
@@ -134,28 +100,28 @@ void main() {
       });
 
       test('false when the idle Explorer is in a different province', () {
-        final game = _supplierGame(
-          extraUnits: [_explorerAt('e1', 'oldWorld|s9')],
+        final game = supplierGame(
+          extraUnits: [supplierProspectExplorerAt('e1', 'oldWorld|s9')],
         );
         expect(
           ownsIdleExplorerColocatedWithMineralEligibleUnprospectedOldWorldFeedstockTile(
             game,
             supplierFeedstockId,
             {'iron'},
-            _tileMapWithIronTerrain(TerrainType.hills),
+            tileMapWithIronTerrain(TerrainType.hills),
           ),
           isFalse,
         );
       });
 
       test('false when the co-located Explorer is busy (currentWork set)', () {
-        final game = _supplierGame(
+        final game = supplierGame(
           extraUnits: [
-            _explorer(
+            supplierProspectExplorer(
               'e1',
               currentWork: const CurrentWork(
                 workTarget: kWorkTargetExplore,
-                tileKey: _newWorldIronTile,
+                tileKey: h8SupplierProspectNewWorldIronTile,
                 totalTurns: 5,
                 remainingTurns: 3,
               ),
@@ -167,35 +133,35 @@ void main() {
             game,
             supplierFeedstockId,
             {'iron'},
-            _tileMapWithIronTerrain(TerrainType.hills),
+            tileMapWithIronTerrain(TerrainType.hills),
           ),
           isFalse,
         );
       });
 
       test('false when the iron tile is already prospected', () {
-        final game = _supplierGame(
-          extraUnits: [_explorer('e1')],
-        ).copyWithSupplierProspected({_supplierIronTile});
+        final game = supplierGame(
+          extraUnits: [supplierProspectExplorer('e1')],
+        ).copyWithSupplierProspected({h8SupplierProspectIronTile});
         expect(
           ownsIdleExplorerColocatedWithMineralEligibleUnprospectedOldWorldFeedstockTile(
             game,
             supplierFeedstockId,
             {'iron'},
-            _tileMapWithIronTerrain(TerrainType.hills),
+            tileMapWithIronTerrain(TerrainType.hills),
           ),
           isFalse,
         );
       });
 
       test('false for an empty feedstock set (negative control)', () {
-        final game = _supplierGame(extraUnits: [_explorer('e1')]);
+        final game = supplierGame(extraUnits: [supplierProspectExplorer('e1')]);
         expect(
           ownsIdleExplorerColocatedWithMineralEligibleUnprospectedOldWorldFeedstockTile(
             game,
             supplierFeedstockId,
             const <String>{},
-            _tileMapWithIronTerrain(TerrainType.hills),
+            tileMapWithIronTerrain(TerrainType.hills),
           ),
           isFalse,
         );
