+++ packages/colonizethis_ai_contracts/test/full_ai_civilian_work_ow_feedstock_reservation_test.dart
@@ -4,7 +4,7 @@ import 'package:colonizethis_logic/colonizethis_logic.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_test/test.dart';
 
-import 'full_ai_civilian_work_supplier_feedstock_extraction_fixtures.dart';
+import 'support/full_ai_civilian_work_supplier_feedstock_extraction_fixtures.dart';
 
 // Refs #2847 § H8-extraction supplier Old World feedstock unit reservation.
 //
