+++ packages/colonizethis_ai_contracts/test/full_ai_civilian_work_seller_feedstock_acquisition_target_among_acquirable_test.dart
@@ -1,8 +1,8 @@
-import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_ai_contracts/colonizethis_ai_contracts.dart';
 import 'package:colonizethis_logic/colonizethis_logic.dart';
-import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_test/test.dart';
+
+import 'support/h8_flagged_seller_game.dart';
 
 // Refs #2847 § H8-extraction seller feedstock-tile acquisition target
 // intersection. Reuses the below-quota zero-NW lock-recovery seller fixture from
@@ -10,50 +10,6 @@ import 'package:colonizethis_test/test.dart';
 // feedstock tile, so the improvement-input gate is active and it needs
 // `lumber` / `castIron`), then exercises the intersection of the topology-free
 // feedstock candidate list with a caller-supplied acquirable province set.
-const _sellerId = 'gp1';
-
-const _grainTile = 'oldWorld|p0|0|0';
-const _woolTile = 'oldWorld|p0|2|0';
-
-Game _flaggedSellerGame({
-  Map<String, String> resourceByTileKey = const {
-    _grainTile: 'grain',
-    _woolTile: 'wool',
-  },
-  List<Province> extraOldWorld = const [],
-  TileMapState? tileState,
-}) {
-  final sellerProvinces = List.generate(
-    5,
-    (i) => Province(
-      id: 'oldWorld|p$i',
-      regionId: kRegionOldWorld,
-      ownerId: _sellerId,
-    ),
-  );
-  return Game(
-    id: 'g',
-    worldState: WorldState(
-      turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 1),
-      oldWorld: RegionData(provinces: [...sellerProvinces, ...extraOldWorld]),
-      newWorld: const RegionData(provinces: []),
-      resourceByTileKey: resourceByTileKey,
-      tileState: tileState ?? TileMapState(),
-    ),
-    players: [
-      Player(
-        id: _sellerId,
-        displayName: 'Seller',
-        isHuman: false,
-        treasury: cheapestRegimentBuildTreasuryCost(),
-        stockpile: const Stockpile(),
-      ),
-    ],
-  );
-}
-
-Province _tribeProvince(String id) =>
-    Province(id: id, regionId: kRegionOldWorld, ownerId: 'tribe1');
 
 void main() {
   group(
@@ -64,32 +20,32 @@ void main() {
         'returns only the feedstock candidates present in the acquirable set, '
         'in ascending province-id order',
         () {
-          final game = _flaggedSellerGame(
+          final game = flaggedSellerGame(
             resourceByTileKey: const {
-              _grainTile: 'grain',
-              _woolTile: 'wool',
+              h8FlaggedSellerGrainTile: 'grain',
+              h8FlaggedSellerWoolTile: 'wool',
               'oldWorld|t1|0|0': 'timber',
               'oldWorld|t2|0|0': 'iron',
               'oldWorld|t3|0|0': 'timber',
             },
             extraOldWorld: [
-              _tribeProvince('oldWorld|t1'),
-              _tribeProvince('oldWorld|t2'),
-              _tribeProvince('oldWorld|t3'),
+              h8TribeProvince('oldWorld|t1'),
+              h8TribeProvince('oldWorld|t2'),
+              h8TribeProvince('oldWorld|t3'),
             ],
           );
           // Precondition: all three feedstock provinces are candidates.
           expect(
             sellerFeedstockTileAcquisitionTargetProvinceIdsSorted(
               game,
-              _sellerId,
+              h8FlaggedSellerId,
             ),
             equals(<String>['oldWorld|t1', 'oldWorld|t2', 'oldWorld|t3']),
           );
           expect(
             sellerFeedstockTileAcquisitionTargetsAmongAcquirable(
               game,
-              _sellerId,
+              h8FlaggedSellerId,
               const {'oldWorld|t2', 'oldWorld|t3', 'oldWorld|t9'},
             ),
             equals(<String>['oldWorld|t2', 'oldWorld|t3']),
@@ -101,22 +57,22 @@ void main() {
         'preserves ascending province-id order regardless of acquirable-set '
         'iteration order',
         () {
-          final game = _flaggedSellerGame(
+          final game = flaggedSellerGame(
             resourceByTileKey: const {
-              _grainTile: 'grain',
-              _woolTile: 'wool',
+              h8FlaggedSellerGrainTile: 'grain',
+              h8FlaggedSellerWoolTile: 'wool',
               'oldWorld|t2|0|0': 'timber',
               'oldWorld|t1|0|0': 'iron',
             },
             extraOldWorld: [
-              _tribeProvince('oldWorld|t2'),
-              _tribeProvince('oldWorld|t1'),
+              h8TribeProvince('oldWorld|t2'),
+              h8TribeProvince('oldWorld|t1'),
             ],
           );
           expect(
             sellerFeedstockTileAcquisitionTargetsAmongAcquirable(
               game,
-              _sellerId,
+              h8FlaggedSellerId,
               // Insertion order deliberately descending.
               const {'oldWorld|t2', 'oldWorld|t1'},
             ),
@@ -126,18 +82,18 @@ void main() {
       );
 
       test('returns empty when the acquirable set is disjoint from candidates', () {
-        final game = _flaggedSellerGame(
+        final game = flaggedSellerGame(
           resourceByTileKey: const {
-            _grainTile: 'grain',
-            _woolTile: 'wool',
+            h8FlaggedSellerGrainTile: 'grain',
+            h8FlaggedSellerWoolTile: 'wool',
             'oldWorld|t1|0|0': 'timber',
           },
-          extraOldWorld: [_tribeProvince('oldWorld|t1')],
+          extraOldWorld: [h8TribeProvince('oldWorld|t1')],
         );
         expect(
           sellerFeedstockTileAcquisitionTargetsAmongAcquirable(
             game,
-            _sellerId,
+            h8FlaggedSellerId,
             const {'oldWorld|t9', 'oldWorld|t8'},
           ),
           isEmpty,
@@ -145,18 +101,18 @@ void main() {
       });
 
       test('returns empty when the acquirable set is empty', () {
-        final game = _flaggedSellerGame(
+        final game = flaggedSellerGame(
           resourceByTileKey: const {
-            _grainTile: 'grain',
-            _woolTile: 'wool',
+            h8FlaggedSellerGrainTile: 'grain',
+            h8FlaggedSellerWoolTile: 'wool',
             'oldWorld|t1|0|0': 'timber',
           },
-          extraOldWorld: [_tribeProvince('oldWorld|t1')],
+          extraOldWorld: [h8TribeProvince('oldWorld|t1')],
         );
         expect(
           sellerFeedstockTileAcquisitionTargetsAmongAcquirable(
             game,
-            _sellerId,
+            h8FlaggedSellerId,
             const <String>{},
           ),
           isEmpty,
@@ -169,23 +125,26 @@ void main() {
         () {
           // Seller owns an unimproved timber tile (oldWorld|p0|1|0), so the
           // routing gate covers it and the detector is false.
-          final game = _flaggedSellerGame(
+          final game = flaggedSellerGame(
             resourceByTileKey: const {
-              _grainTile: 'grain',
-              _woolTile: 'wool',
+              h8FlaggedSellerGrainTile: 'grain',
+              h8FlaggedSellerWoolTile: 'wool',
               'oldWorld|p0|1|0': 'timber',
               'oldWorld|t1|0|0': 'timber',
             },
-            extraOldWorld: [_tribeProvince('oldWorld|t1')],
+            extraOldWorld: [h8TribeProvince('oldWorld|t1')],
           );
           expect(
-            sellerNeedsImprovementInputFeedstockTileAcquisition(game, _sellerId),
+            sellerNeedsImprovementInputFeedstockTileAcquisition(
+              game,
+              h8FlaggedSellerId,
+            ),
             isFalse,
           );
           expect(
             sellerFeedstockTileAcquisitionTargetsAmongAcquirable(
               game,
-              _sellerId,
+              h8FlaggedSellerId,
               const {'oldWorld|t1'},
             ),
             isEmpty,
@@ -194,27 +153,27 @@ void main() {
       );
 
       test('evaluation is deterministic', () {
-        final game = _flaggedSellerGame(
+        final game = flaggedSellerGame(
           resourceByTileKey: const {
-            _grainTile: 'grain',
-            _woolTile: 'wool',
+            h8FlaggedSellerGrainTile: 'grain',
+            h8FlaggedSellerWoolTile: 'wool',
             'oldWorld|t2|0|0': 'timber',
             'oldWorld|t1|0|0': 'iron',
           },
           extraOldWorld: [
-            _tribeProvince('oldWorld|t2'),
-            _tribeProvince('oldWorld|t1'),
+            h8TribeProvince('oldWorld|t2'),
+            h8TribeProvince('oldWorld|t1'),
           ],
         );
         const acquirable = {'oldWorld|t1', 'oldWorld|t2'};
         final a = sellerFeedstockTileAcquisitionTargetsAmongAcquirable(
           game,
-          _sellerId,
+          h8FlaggedSellerId,
           acquirable,
         );
         final b = sellerFeedstockTileAcquisitionTargetsAmongAcquirable(
           game,
-          _sellerId,
+          h8FlaggedSellerId,
           acquirable,
         );
         expect(a, equals(b));
