+++ packages/colonizethis_ai_contracts/test/full_ai_civilian_work_seller_feedstock_acquisition_target_test.dart
@@ -1,8 +1,9 @@
-import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_ai_contracts/colonizethis_ai_contracts.dart';
+import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_logic/colonizethis_logic.dart';
-import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_test/test.dart';
+
+import 'support/h8_flagged_seller_game.dart';
 
 // Refs #2847 § H8-extraction seller feedstock-tile acquisition target selection.
 // Builds the same below-quota zero-NW lock-recovery seller fixture the detection
@@ -10,51 +11,6 @@ import 'package:colonizethis_test/test.dart';
 // so the improvement-input gate is active and it needs `lumber` / `castIron`),
 // then adds non-seller-owned provinces hosting `timber` / `iron` feedstock so the
 // selection contract has acquisition candidates to enumerate.
-const _sellerId = 'gp1';
-
-const _grainTile = 'oldWorld|p0|0|0';
-const _woolTile = 'oldWorld|p0|2|0';
-
-Game _flaggedSellerGame({
-  Map<String, String> resourceByTileKey = const {
-    _grainTile: 'grain',
-    _woolTile: 'wool',
-  },
-  List<Province> extraOldWorld = const [],
-  List<Province> extraNewWorld = const [],
-  TileMapState? tileState,
-}) {
-  final sellerProvinces = List.generate(
-    5,
-    (i) => Province(
-      id: 'oldWorld|p$i',
-      regionId: kRegionOldWorld,
-      ownerId: _sellerId,
-    ),
-  );
-  return Game(
-    id: 'g',
-    worldState: WorldState(
-      turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 1),
-      oldWorld: RegionData(provinces: [...sellerProvinces, ...extraOldWorld]),
-      newWorld: RegionData(provinces: extraNewWorld),
-      resourceByTileKey: resourceByTileKey,
-      tileState: tileState ?? TileMapState(),
-    ),
-    players: [
-      Player(
-        id: _sellerId,
-        displayName: 'Seller',
-        isHuman: false,
-        treasury: cheapestRegimentBuildTreasuryCost(),
-        stockpile: const Stockpile(),
-      ),
-    ],
-  );
-}
-
-Province _tribeProvince(String id, {String region = kRegionOldWorld}) =>
-    Province(id: id, regionId: region, ownerId: 'tribe1');
 
 void main() {
   group(
@@ -65,23 +21,26 @@ void main() {
         'returns the single acquirable Old World province hosting timber '
         'feedstock when the seller is flagged',
         () {
-          final game = _flaggedSellerGame(
+          final game = flaggedSellerGame(
             resourceByTileKey: const {
-              _grainTile: 'grain',
-              _woolTile: 'wool',
+              h8FlaggedSellerGrainTile: 'grain',
+              h8FlaggedSellerWoolTile: 'wool',
               'oldWorld|t1|0|0': 'timber',
             },
-            extraOldWorld: [_tribeProvince('oldWorld|t1')],
+            extraOldWorld: [h8TribeProvince('oldWorld|t1')],
           );
           // Precondition: the acquisition residual is active.
           expect(
-            sellerNeedsImprovementInputFeedstockTileAcquisition(game, _sellerId),
+            sellerNeedsImprovementInputFeedstockTileAcquisition(
+              game,
+              h8FlaggedSellerId,
+            ),
             isTrue,
           );
           expect(
             sellerFeedstockTileAcquisitionTargetProvinceIdsSorted(
               game,
-              _sellerId,
+              h8FlaggedSellerId,
             ),
             equals(<String>['oldWorld|t1']),
           );
@@ -92,22 +51,22 @@ void main() {
         'returns every feedstock-bearing acquirable province sorted ascending '
         'by province id',
         () {
-          final game = _flaggedSellerGame(
+          final game = flaggedSellerGame(
             resourceByTileKey: const {
-              _grainTile: 'grain',
-              _woolTile: 'wool',
+              h8FlaggedSellerGrainTile: 'grain',
+              h8FlaggedSellerWoolTile: 'wool',
               'oldWorld|t2|0|0': 'timber',
               'oldWorld|t1|0|0': 'iron',
             },
             extraOldWorld: [
-              _tribeProvince('oldWorld|t2'),
-              _tribeProvince('oldWorld|t1'),
+              h8TribeProvince('oldWorld|t2'),
+              h8TribeProvince('oldWorld|t1'),
             ],
           );
           expect(
             sellerFeedstockTileAcquisitionTargetProvinceIdsSorted(
               game,
-              _sellerId,
+              h8FlaggedSellerId,
             ),
             equals(<String>['oldWorld|t1', 'oldWorld|t2']),
           );
@@ -118,25 +77,28 @@ void main() {
         'returns empty when the acquisition residual is inactive (seller owns '
         'an unimproved timber tile)',
         () {
-          final game = _flaggedSellerGame(
+          final game = flaggedSellerGame(
             resourceByTileKey: const {
-              _grainTile: 'grain',
-              _woolTile: 'wool',
+              h8FlaggedSellerGrainTile: 'grain',
+              h8FlaggedSellerWoolTile: 'wool',
               'oldWorld|p0|1|0': 'timber',
               'oldWorld|t1|0|0': 'timber',
             },
-            extraOldWorld: [_tribeProvince('oldWorld|t1')],
+            extraOldWorld: [h8TribeProvince('oldWorld|t1')],
           );
           // The routing gate covers the seller's own unimproved timber tile, so
           // the detector is false and no acquisition target is offered.
           expect(
-            sellerNeedsImprovementInputFeedstockTileAcquisition(game, _sellerId),
+            sellerNeedsImprovementInputFeedstockTileAcquisition(
+              game,
+              h8FlaggedSellerId,
+            ),
             isFalse,
           );
           expect(
             sellerFeedstockTileAcquisitionTargetProvinceIdsSorted(
               game,
-              _sellerId,
+              h8FlaggedSellerId,
             ),
             isEmpty,
           );
@@ -147,24 +109,27 @@ void main() {
         'excludes New World feedstock provinces (cannot close the Old World '
         'turn-100 gate)',
         () {
-          final game = _flaggedSellerGame(
+          final game = flaggedSellerGame(
             resourceByTileKey: const {
-              _grainTile: 'grain',
-              _woolTile: 'wool',
+              h8FlaggedSellerGrainTile: 'grain',
+              h8FlaggedSellerWoolTile: 'wool',
               'newWorld|n1|0|0': 'timber',
             },
             extraNewWorld: [
-              _tribeProvince('newWorld|n1', region: kRegionNewWorld),
+              h8TribeProvince('newWorld|n1', region: kRegionNewWorld),
             ],
           );
           expect(
-            sellerNeedsImprovementInputFeedstockTileAcquisition(game, _sellerId),
+            sellerNeedsImprovementInputFeedstockTileAcquisition(
+              game,
+              h8FlaggedSellerId,
+            ),
             isTrue,
           );
           expect(
             sellerFeedstockTileAcquisitionTargetProvinceIdsSorted(
               game,
-              _sellerId,
+              h8FlaggedSellerId,
             ),
             isEmpty,
           );
@@ -174,22 +139,25 @@ void main() {
       test(
         'excludes acquirable provinces that host only non-feedstock resources',
         () {
-          final game = _flaggedSellerGame(
+          final game = flaggedSellerGame(
             resourceByTileKey: const {
-              _grainTile: 'grain',
-              _woolTile: 'wool',
+              h8FlaggedSellerGrainTile: 'grain',
+              h8FlaggedSellerWoolTile: 'wool',
               'oldWorld|t1|0|0': 'grain',
             },
-            extraOldWorld: [_tribeProvince('oldWorld|t1')],
+            extraOldWorld: [h8TribeProvince('oldWorld|t1')],
           );
           expect(
-            sellerNeedsImprovementInputFeedstockTileAcquisition(game, _sellerId),
+            sellerNeedsImprovementInputFeedstockTileAcquisition(
+              game,
+              h8FlaggedSellerId,
+            ),
             isTrue,
           );
           expect(
             sellerFeedstockTileAcquisitionTargetProvinceIdsSorted(
               game,
-              _sellerId,
+              h8FlaggedSellerId,
             ),
             isEmpty,
           );
@@ -197,25 +165,25 @@ void main() {
       );
 
       test('evaluation is deterministic', () {
-        final game = _flaggedSellerGame(
+        final game = flaggedSellerGame(
           resourceByTileKey: const {
-            _grainTile: 'grain',
-            _woolTile: 'wool',
+            h8FlaggedSellerGrainTile: 'grain',
+            h8FlaggedSellerWoolTile: 'wool',
             'oldWorld|t2|0|0': 'timber',
             'oldWorld|t1|0|0': 'iron',
           },
           extraOldWorld: [
-            _tribeProvince('oldWorld|t2'),
-            _tribeProvince('oldWorld|t1'),
+            h8TribeProvince('oldWorld|t2'),
+            h8TribeProvince('oldWorld|t1'),
           ],
         );
         final a = sellerFeedstockTileAcquisitionTargetProvinceIdsSorted(
           game,
-          _sellerId,
+          h8FlaggedSellerId,
         );
         final b = sellerFeedstockTileAcquisitionTargetProvinceIdsSorted(
           game,
-          _sellerId,
+          h8FlaggedSellerId,
         );
         expect(a, equals(b));
       });
